Change reps() and deps() for safe

+++ cpp.cpp
@@ -10,11 +10,11 @@ typedef queue<int> qi;
 typedef stack<int> si;
 typedef deque<int> dqi;
 
-#define reps(i, a, b) for (int i = a; i < b; ++i)
+#define reps(i, a, b) for (int i = (a); i < (int)(b); ++i)
 #define rep(i, n) reps(i, 0, n)
 #define REPS(i, a, b) reps(i, a, b)
 #define REP(i, n) rep(i, n)
-#define deps(i, a, b) for (int i = a; i >= b; --i)
+#define deps(i, a, b) for (int i = (a); i >= (int)(b); --i)
 #define dep(i, n) deps(i, n, 0)
 #define DEPS(i, a, b) deps(i, a, b)
 #define DEP(i, n) dep(i, n)
